MediumArticles JetpackComposeTabs changes following moving of class
